issue #35: move classes around to fix missing file error
