+++ classes/local/renderer/combined/renderer.php
@@ -22,7 +22,7 @@
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-namespace quiz_answersheets\output\combined;
+namespace quiz_answersheets\local\renderer\combined;
 
 use coding_exception;
 use html_writer;
