+++ classes/local/renderer/match/renderer.php
@@ -22,7 +22,7 @@
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-namespace quiz_answersheets\output\match;
+namespace quiz_answersheets\local\renderer\match;
 
 use html_writer;
 use question_attempt;
