+++ classes/local/renderer/stack/renderer.php
@@ -22,7 +22,7 @@
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-namespace quiz_answersheets\output\stack;
+namespace quiz_answersheets\local\renderer\stack;
 
 defined('MOODLE_INTERNAL') || die();
