+++ classes/local/renderer/truefalse/renderer.php
@@ -22,7 +22,7 @@
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-namespace quiz_answersheets\output\truefalse;
+namespace quiz_answersheets\local\renderer\truefalse;
 
 use html_writer;
 use question_attempt;
